refactor code to meet coding rules
