+++ server/src/Auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { AuthService } from './auth.service';
 import { Module } from '@nestjs/common';
-import { UserModule } from 'src/User/user.module';
+import { UserModule } from 'src/user/user.module';
 import { PassportModule } from '@nestjs/passport';
 import { LocalStrategy } from './local.strategy';
 import { JwtModule } from '@nestjs/jwt';
@@ -8,7 +8,7 @@ import { jwtConstants } from './constants';
 import { JwtStrategy } from './jwt.strategy';
 import { MailModule } from 'src/mail/mail.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from 'src/User/user.entity';
+import { User } from 'src/user/entities/user.entity';
 
 @Module({
   imports: [
