+++ server/src/Auth/auth.service.ts
@@ -5,11 +5,11 @@ import {
   HttpStatus,
   Inject
 } from '@nestjs/common';
-import { User } from 'src/User/user.entity';
-import { UserService } from 'src/User/user.service';
+import { User } from 'src/user/entities/user.entity';
+import { UserService } from 'src/user/user.service';
 import { JwtService } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
-import { SignUpDto } from 'src/Dto/user.dto';
+import { SignUpDto } from 'src/dto/user.dto';
 
 import { MailService } from 'src/mail/mail.service';
 import { sign } from 'crypto';
