+++ server/src/Room/room.module.ts
@@ -2,9 +2,9 @@ import { RoomController } from './room.controller';
 import { RoomService } from './room.service';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Room } from './room.entity';
-import { User } from 'src/User/user.entity';
-import { Question } from 'src/Question/entities/question.entity';
+import { Room } from './entities/room.entity';
+import { User } from 'src/user/entities/user.entity';
+import { Question } from 'src/question/entities/question.entity';
 import { Answer } from 'src/answer/entities/answer.entity';
 
 @Module({
