+++ server/src/answer/answer.controller.ts
@@ -12,7 +12,7 @@ import {
 import { AnswerService } from './answer.service';
 import { CreateAnswerDto } from './dto/create-answer.dto';
 import { UpdateAnswerDto } from './dto/update-answer.dto';
-import { JwtAuthGuard } from '../Auth/jwt-auth.guard';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 
 @Controller('api/answers')
 export class AnswerController {
