+++ server/src/app.module.ts
@@ -1,13 +1,13 @@
-import { RoomModule } from './Room/room.module';
-import { AuthModule } from './Auth/auth.module';
-import { UserModule } from './User/user.module';
+import { RoomModule } from './room/room.module';
+import { AuthModule } from './auth/auth.module';
+import { UserModule } from './user/user.module';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
 import { MailModule } from './mail/mail.module';
-import { QuestionModule } from './Question/question.module';
+import { QuestionModule } from './question/question.module';
 import { AnswerModule } from './answer/answer.module';
 import ormConfig from './config/orm.config';
 import ormConfigProd from './config/orm.config.prod';
