+++ server/src/tag/entities/tag.entity.ts
@@ -7,7 +7,7 @@ import {
   ManyToMany,
   JoinTable
 } from 'typeorm';
-import { Question } from 'src/Question/entities/question.entity';
+import { Question } from 'src/question/entities/question.entity';
 
 @Entity()
 export class Tag {
